Move testing source functions out of sm_sir
