+++ autumn/core/inputs/testing/eur_testing_data.py
@@ -1,9 +1,19 @@
-import os
+from typing import Tuple, Optional
 import pandas as pd
+import os
 from datetime import datetime
 
+from autumn.settings import COVID_BASE_AGEGROUPS
+from autumn.core.utils.utils import apply_moving_average
+from autumn.model_features.curve import scale_up_function
+from autumn.core.inputs.covid_au.queries import get_vic_testing_numbers
+from autumn.core.inputs.covid_phl.queries import get_phl_subregion_testing_numbers
+from autumn.core.inputs.covid_lka.queries import get_lka_testing_numbers
+from autumn.core.inputs.covid_mmr.queries import get_mmr_testing_numbers
+from autumn.core.inputs.covid_bgd.queries import get_coxs_bazar_testing_numbers
+from autumn.core.inputs.owid.queries import get_international_testing_numbers
+from autumn.core.inputs.covid_btn.queries import get_btn_testing_numbers
 from autumn.settings.folders import INPUT_DATA_PATH
-
 from autumn.settings.constants import COVID_BASE_DATETIME
 
 base_dir = os.path.dirname(os.path.abspath(os.curdir))
@@ -66,3 +76,52 @@ def convert_weekly_total_to_daily_numbers(year_week, n_tests_per_week):
         test_numbers += [week_val / 7.0 for _ in range(7)]
 
     return test_dates, test_numbers
+
+
+def get_testing_numbers_for_region(
+    country_iso3: str, 
+    subregion: Optional[str]
+) -> Tuple[list, list]:
+    """
+    Use the appropriate function to retrieve the testing numbers applicable to the region being modelled.
+    Functions are taken from the autumn input tools module, as above.
+
+    Args:
+        country_iso3: ISO3 code for the country being simulated
+        subregion: Name of the country's subregion being considered or None if it is a national level
+    Return:
+        The testing data for the country
+    """
+
+    subregion = subregion or False
+
+    if country_iso3 == "AUS":
+        test_dates, test_values = get_vic_testing_numbers()
+    elif country_iso3 == "PHL":
+        phl_region = subregion.lower() if subregion else "philippines"
+        test_dates, test_values = get_phl_subregion_testing_numbers(phl_region)
+    elif subregion == "Sabah":
+        test_dates, test_values = get_international_testing_numbers(country_iso3)
+    elif country_iso3 == "GBR":
+        test_dates, test_values = get_uk_testing_numbers()
+    elif country_iso3 in {"BEL", "ITA", "SWE", "FRA", "ESP"}:
+        test_dates, test_values = get_eu_testing_numbers(country_iso3)
+    elif country_iso3 == "LKA":
+        test_dates, test_values = get_lka_testing_numbers()
+    elif country_iso3 == "MMR":
+        test_dates, test_values = get_mmr_testing_numbers()
+    elif country_iso3 == "BGD" and subregion == "FDMN":
+        test_dates, test_values = get_coxs_bazar_testing_numbers()
+    elif country_iso3 == "BTN":
+        test_dates, test_values = get_btn_testing_numbers(subregion)
+
+    else:
+        test_dates, test_values = get_international_testing_numbers(country_iso3)
+
+    # Check data and return
+    msg = "Length of test dates and test values are not equal"
+    assert len(test_dates) == len(test_values), msg
+    test_df = pd.Series(test_values, index=test_dates)
+    msg = "Negative test values present"
+    assert (test_df >= 0).all()
+    return test_df
