+++ autumn/models/sm_sir/detection.py
@@ -1,69 +1,11 @@
 from typing import Tuple, Callable, Optional
 import numpy as np
-import pandas as pd
 
+from autumn.core.inputs.testing.eur_testing_data import get_testing_numbers_for_region
 from .parameters import TestingToDetection, Population
 from autumn.core.inputs import get_population_by_agegroup
 from autumn.settings import COVID_BASE_AGEGROUPS
-from autumn.core.utils.utils import apply_moving_average
 from autumn.model_features.curve import scale_up_function
-from autumn.core.inputs.testing.eur_testing_data import get_uk_testing_numbers, get_eu_testing_numbers
-from autumn.core.inputs.covid_au.queries import get_vic_testing_numbers
-from autumn.core.inputs.covid_phl.queries import get_phl_subregion_testing_numbers
-from autumn.core.inputs.covid_lka.queries import get_lka_testing_numbers
-from autumn.core.inputs.covid_mmr.queries import get_mmr_testing_numbers
-from autumn.core.inputs.covid_bgd.queries import get_coxs_bazar_testing_numbers
-from autumn.core.inputs.owid.queries import get_international_testing_numbers
-from autumn.core.inputs.covid_btn.queries import get_btn_testing_numbers
-
-
-def get_testing_numbers_for_region(
-    country_iso3: str, 
-    subregion: Optional[str]
-) -> Tuple[list, list]:
-    """
-    Use the appropriate function to retrieve the testing numbers applicable to the region being modelled.
-    Functions are taken from the autumn input tools module, as above.
-
-    Args:
-        country_iso3: ISO3 code for the country being simulated
-        subregion: Name of the country's subregion being considered or None if it is a national level
-    Return:
-        The testing data for the country
-    """
-
-    subregion = subregion or False
-
-    if country_iso3 == "AUS":
-        test_dates, test_values = get_vic_testing_numbers()
-    elif country_iso3 == "PHL":
-        phl_region = subregion.lower() if subregion else "philippines"
-        test_dates, test_values = get_phl_subregion_testing_numbers(phl_region)
-    elif subregion == "Sabah":
-        test_dates, test_values = get_international_testing_numbers(country_iso3)
-    elif country_iso3 == "GBR":
-        test_dates, test_values = get_uk_testing_numbers()
-    elif country_iso3 in {"BEL", "ITA", "SWE", "FRA", "ESP"}:
-        test_dates, test_values = get_eu_testing_numbers(country_iso3)
-    elif country_iso3 == "LKA":
-        test_dates, test_values = get_lka_testing_numbers()
-    elif country_iso3 == "MMR":
-        test_dates, test_values = get_mmr_testing_numbers()
-    elif country_iso3 == "BGD" and subregion == "FDMN":
-        test_dates, test_values = get_coxs_bazar_testing_numbers()
-    elif country_iso3 == "BTN":
-        test_dates, test_values = get_btn_testing_numbers(subregion)
-
-    else:
-        test_dates, test_values = get_international_testing_numbers(country_iso3)
-
-    # Check data and return
-    msg = "Length of test dates and test values are not equal"
-    assert len(test_dates) == len(test_values), msg
-    test_df = pd.Series(test_values, index=test_dates)
-    msg = "Negative test values present"
-    assert (test_df >= 0).all()
-    return test_df
 
 
 def create_cdr_function(assumed_tests: int, assumed_cdr: float, floor_value: float=0.) -> Callable:
@@ -125,15 +67,9 @@ def find_cdr_function_from_test_data(
     cdr_from_tests_func = create_cdr_function(*cdr_test_params)
 
     # Get the final CDR function
-    cdr_function = scale_up_function(
-        per_capita_tests_df.index,
-        per_capita_tests_df.apply(cdr_from_tests_func),
-        smoothness=0.2,
-        method=4,
-        bound_low=0.0,
-    )
-
-    return cdr_function
+    times = per_capita_tests_df.index
+    values = per_capita_tests_df.apply(cdr_from_tests_func)
+    return scale_up_function(times, values, smoothness=0.2, method=4, bound_low=0.0)
 
 
 def get_cdr_func(
